+++ x11/x11_icon.h
@@ -14,7 +14,7 @@ class X11Icon final : public Icon
 {
 public:
     X11Icon(int width, int height, X11Widget *parent);
-    X11Icon(int width, int height, X11Widget *parent, const uint32 *argb_data, uint32 bg_color);
+    X11Icon(int width, int height, X11Widget *parent, const unsigned long *argb_data, uint32 bg_color);
     X11Icon(const X11Icon &other);
     ~X11Icon();
     X11Icon &operator=(const X11Icon &other);
